source/requirements.txt
Django==2.2.4
django-appconf==1.0.4
django-user-accounts==3.0.0
django-widget-tweaks==1.4.8
djangorestframework==3.11.0
mysqlclient==1.4.6
Pillow==7.1.1
PyMySQL==0.9.3
pytz==2019.3
sqlparse==0.3.1
